refac[Geração de parcelas]: alterado tipo de variável e extraído controle de data para outra classe

+++ src/main/java/br/edu/ufape/agiota/dtos/AprovarEmprestimoDTO.java
@@ -24,7 +24,7 @@ public class AprovarEmprestimoDTO {
 
     @NotNull
     @Positive
-    private double periodoParcelas;
+    private int periodoParcelas;
 
     public void aprovar(Emprestimo emprestimo, double taxaDeJuros) {
         emprestimo.setDataEmprestimo(getDataEmprestimo());
